==> 4. cougar analyses/4.  cougar conflict scripts/1. cougar increase data.R <==

# Prep Wolf Public Opinion Data: ------------------------------------------
  ## Here we bring in the wolf increase public response data and crop it to the BH Subwatershed region


# Load packages: ----------------------------------------------------------
library(terra)



# Bring in Data: ----------------------------------------------------------
